feat: Implement ReviewerController (feature/reviewer)
This commit implements the ReviewerController interface, providing
endpoints for registering, retrieving, and updating reviewer
information.  It includes the necessary logic for handling
requests and interacting with the UserService.
